requirements.txt
python-binance
TA-lib
numpy
websocket_client